pipeline fix (#1278)
* pipeline fix

* pipeline fix

* pipeline fix

* review comment
