+++ config/development.yaml
@@ -21,11 +21,11 @@ services:
   citizen:
     uploadDocsEmail: UPLOAD_DOCUMENTS_EMAIL
   case:
-    url: 'http://ccd-data-store-api-aat.service.core-compute-aat.internal'
+    url: 'https://ccd-data-store-api-prl-ccd-definitions-pr-1909.preview.platform.hmcts.net'
   cos:
-    url: 'http://prl-cos-aat.service.core-compute-aat.internal'
+    url: 'https://prl-cos-pr-2073.preview.platform.hmcts.net'
   documentManagement:
-    url: 'http://ccd-case-document-am-api-aat.service.core-compute-aat.internal'
+    url: 'https://ccd-case-document-am-api-prl-ccd-definitions-pr-1909.preview.platform.hmcts.net'
   health:
   timeout: 5000
   deadline: 15000
