+++ sonar-project.properties
@@ -6,7 +6,7 @@ sonar.tests=src/test/
 sonar.test.exclusions=src/**/*.njk
 sonar.typescript.lcov.reportPaths=coverage/lcov.info
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
-sonar.exclusions=src/**/*.test.ts,src/main/steps/tasklistresponse/proceedings/proceedingUtils.ts,src/main/steps/tasklistresponse/proceedings/ProceedingDetailsMapper.ts,src/main/steps/tasklistresponse/proceedings/mainUtils.ts,src/main/steps/tasklistresponse/proceedings/documentUpload/postController.ts,src/main/steps/tasklistresponse/responseCaseSequence.ts,src/main/app/document/DocumentManagementController.ts,src/main/steps/common/confirm-contact-details/checkanswers/ConfirmContactDetailsPostController.ts,src/main/app/controller/GetCaseController.ts,src/main/app/controller/PostController.ts,src/main/app/controller/RespondentSubmitResponseController.ts,src/main/steps/common/confirm-contact-details/checkanswers/ContactDetailsMapper.ts,src/main/steps/common/confirm-contact-details/checkanswers/ConfirmContactDetailsGetController.ts,src/main/steps/tasklistresponse/start/tasklist.ts,src/main/steps/tasklistresponse/tasklistresponseCaseSequence.ts,src/main/app/document/DocumentManagementClient.ts,src/main/steps/respondent/task-list/tasklist.ts,**/__mocks__/**,src/**/*content.ts,src/**/definition.ts,src/**/steps/constants.ts,src/**/case/CaseApi.ts,src/main/steps/respondent/task-list/utils.ts,src/main/app/case/C100CaseApi.ts,src/main/steps/c100-rebuild/check-your-answers/mainUtil.ts,src/main/steps/tasklistresponse/allegations-of-harm-and-violence/check-your-answers/mainUtil.ts,src/main/steps/tasklistresponse/allegations-of-harm-and-violence/SafetyConcernsMapper.ts,src/main/steps/tasklistresponse/allegations-of-harm-and-violence/check-your-answers/util/safetyConcerns.util.ts,src/main/steps/common/support-you-need-during-case/summary/utils.ts,src/main/steps/common/summary/utils.ts,src/main/steps/common/dateformatter.ts,src/main/steps/tasklistresponse/allegations-of-harm-and-violence/check-your-answers/helpers/satetyConcernHelper.ts,src/main/steps/applicant/yourdocuments/alldocuments/alldocuments/tasklistalldocuments.ts, src/main/steps/c100-rebuild/other-proceedings/documentUpload/getController.ts,src/main/steps/tasklistresponse/proceedings/documentUpload/getController.ts
+sonar.exclusions=src/**/*.test.ts,src/main/steps/tasklistresponse/proceedings/proceedingUtils.ts,src/main/steps/tasklistresponse/proceedings/ProceedingDetailsMapper.ts,src/main/steps/tasklistresponse/proceedings/mainUtils.ts,src/main/steps/tasklistresponse/proceedings/documentUpload/postController.ts,src/main/steps/tasklistresponse/responseCaseSequence.ts,src/main/app/document/DocumentManagementController.ts,src/main/steps/common/confirm-contact-details/checkanswers/ConfirmContactDetailsPostController.ts,src/main/app/controller/GetCaseController.ts,src/main/app/controller/PostController.ts,src/main/app/controller/RespondentSubmitResponseController.ts,src/main/steps/common/confirm-contact-details/checkanswers/ContactDetailsMapper.ts,src/main/steps/common/confirm-contact-details/checkanswers/ConfirmContactDetailsGetController.ts,src/main/steps/tasklistresponse/start/tasklist.ts,src/main/steps/tasklistresponse/tasklistresponseCaseSequence.ts,src/main/app/document/DocumentManagementClient.ts,src/main/steps/respondent/task-list/tasklist.ts,**/__mocks__/**,src/**/*content.ts,src/**/definition.ts,src/**/steps/constants.ts,src/**/case/CaseApi.ts,src/main/steps/respondent/task-list/utils.ts,src/main/app/case/C100CaseApi.ts,src/main/steps/c100-rebuild/check-your-answers/mainUtil.ts,src/main/steps/tasklistresponse/allegations-of-harm-and-violence/check-your-answers/mainUtil.ts,src/main/steps/tasklistresponse/allegations-of-harm-and-violence/SafetyConcernsMapper.ts,src/main/steps/tasklistresponse/allegations-of-harm-and-violence/check-your-answers/util/safetyConcerns.util.ts,src/main/steps/common/support-you-need-during-case/summary/utils.ts,src/main/steps/common/summary/utils.ts,src/main/steps/common/dateformatter.ts,src/main/steps/tasklistresponse/allegations-of-harm-and-violence/check-your-answers/helpers/satetyConcernHelper.ts,src/main/steps/applicant/yourdocuments/alldocuments/alldocuments/tasklistalldocuments.ts
 sonar.organization=hmcts
 sonar.cpd.typescript.minimumLines=250
 sonar.cpd.javascript.minimumLines=250
