+++ src/main/steps/c100-rebuild/other-proceedings/documentUpload/getController.test.ts
@@ -3,6 +3,8 @@ import axios from 'axios';
 import { mockRequest } from '../../../../../test/unit/utils/mockRequest';
 import { mockResponse } from '../../../../../test/unit/utils/mockResponse';
 import { FieldPrefix } from '../../../../app/case/case';
+import { applyParms } from '../../../common/url-parser';
+import { C100_OTHER_PROCEEDINGS_DOCUMENT_UPLOAD } from '../../../urls';
 
 import EmergencyDocumentUpload from './getController';
 
@@ -42,6 +44,7 @@ describe('DocumentUpload Get Controller', () => {
       orderType: 'otherOrder',
       orderId: '1',
     };
+    req.originalUrl = applyParms(C100_OTHER_PROCEEDINGS_DOCUMENT_UPLOAD, { orderType: 'otherOrder', orderId: '1' });
 
     await controller.get(req, res);
 
@@ -148,6 +151,7 @@ describe('DocumentUpload Get Controller', () => {
         orderType: 'otherOrder',
         orderId: '1',
       };
+      req.originalUrl = applyParms(C100_OTHER_PROCEEDINGS_DOCUMENT_UPLOAD, { orderType: 'otherOrder', orderId: '1' });
 
       await controller.get(req, res);
