+++ src/main/steps/c100-rebuild/other-proceedings/documentUpload/postController.ts
@@ -31,8 +31,8 @@ const C100OrderTypeNameMapper = {
   otherOrder: 'Other Order',
 };
 
-// eslint-disable-next-line import/no-unresolved
-import { AnyType } from './getController';
+//eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type AnyType = any;
 
 /* The UploadDocumentController class extends the PostController class and overrides the
 PostDocumentUploader method */
